requirements.txt
asgiref==3.11.0
certifi==2025.11.12
charset-normalizer==3.4.4
coreapi==2.3.3
coreschema==0.0.4
Django==6.0
django-cors-headers==4.9.0
django-filter==25.2
django-silk==5.4.3
djangorestframework==3.16.1
drf-yasg==1.21.11
future==1.0.0
gprof2dot==2025.4.14
idna==3.11
inflection==0.5.1
iniconfig==2.3.0
itypes==1.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
openapi-codec==1.3.2
packaging==25.0
pillow==12.0.0
pluggy==1.6.0
Pygments==2.19.2
pytest==7.4.3
pytest-django==4.11.1
pytz==2025.2
PyYAML==6.0.3
requests==2.32.5
ruamel.yaml==0.18.16
ruamel.yaml.clib==0.2.15
sqlparse==0.5.4
uritemplate==4.2.0
urllib3==2.6.2
